Move folder from integration-test to docs/integration-test
