+++ docs/integration-test/backend/stop.sh
