+++ docs/integration-test/cli-scenario/cli_full_json_test.go
