+++ docs/integration-test/cli-scenario/cli_full_yaml_test.go
