+++ docs/integration-test/cli-scenario/setup.go
