+++ docs/integration-test/conf/store_conf.yaml
