+++ docs/integration-test/go-scenario/go_full_doc_arg_test.go
