+++ docs/integration-test/go-scenario/method_call.go
