+++ docs/integration-test/rest-scenario/echo_call.go
