+++ docs/integration-test/rest-scenario/rest_duplicate_test.go
