+++ docs/integration-test/rest-scenario/rest_full_test.go
