+++ docs/integration-test/test.env
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-export CBLOG_ROOT=$GOPATH/src/github.com/cloud-barista/poc-cicd-ladybug/integration-test
-export CBSTORE_ROOT=$GOPATH/src/github.com/cloud-barista/poc-cicd-ladybug/integration-test
-export APP_ROOT=$GOPATH/src/github.com/cloud-barista/poc-cicd-ladybug/integration-test
+export CBLOG_ROOT=$GOPATH/src/github.com/cloud-barista/poc-cicd-ladybug/docs/integration-test
+export CBSTORE_ROOT=$GOPATH/src/github.com/cloud-barista/poc-cicd-ladybug/docs/integration-test
+export APP_ROOT=$GOPATH/src/github.com/cloud-barista/poc-cicd-ladybug/docs/integration-test
 
 export SPIDER_URL=http://localhost:31024/spider # *** Be careful : port is changed for testing
 export TUMBLEBUG_URL=http://localhost:31323/tumblebug # *** Be careful : port is changed for testing
